backend_services/ai_services/summerization_service/pyproject.toml
[project]
name = "summerization-service"
version = "0.1.0"
description = "Document summarization microservice for Care Connect"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.2.2",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "docling>=2.0.0",
    "httpx>=0.27.0",
]
